requirements.txt
py-cord==2.6.1
supabase==2.10.0
openai==1.51.0
pytesseract==0.3.13
Pillow==10.4.0
matplotlib==3.9.2
python-dotenv==1.0.1
pydantic==2.9.2
aiohttp==3.10.10
python-magic==0.4.27
httpx==0.27.2
